LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-disconnect-expected.txt


FAIL Disconnecting and reconnecting host during async fetch still applies styles. assert_equals: Before disconnection: expected 1 placeholder(s). expected 1 but got 0
